+++ kmath-ast/src/jsMain/kotlin/kscience/kmath/internal/astring/astring.typealises.kt
@@ -1,3 +1,0 @@
-package kscience.kmath.internal.astring
-
-internal typealias Generator = Any